refactor(atomic): move common/interface/item.ts to relevant folder
https://coveord.atlassian.net/browse/KIT-4637
