+++ packages/atomic/src/components/common/item-list/item-list-common.ts
@@ -9,11 +9,11 @@ import type {CommerceRecommendationStore} from '../../commerce/atomic-commerce-r
 import type {InsightStore} from '../../insight/atomic-insight-interface/store';
 import type {RecsStore} from '../../recommendations/atomic-recs-interface/store';
 import type {SearchStore} from '../../search/atomic-search-interface/store';
-import type {AnyItem} from '../interface/item';
 import type {
   ItemDisplayDensity,
   ItemDisplayImageSize,
 } from '../layout/display-options';
+import type {AnyItem} from './unfolded-item';
 
 export const resultComponentClass = 'result-component';
