+++ packages/atomic/src/components/common/item-list/stencil-display-table.tsx
@@ -1,7 +1,7 @@
 // The Lit equivalent of this file is table-layout.ts
 import {FunctionalComponent, VNode, h} from '@stencil/core';
 import {tableElementTagName} from '../../search/atomic-table-result/table-element-utils.js';
-import {AnyItem} from '../interface/item.js';
+import {AnyItem} from './unfolded-item.js';
 import {ItemRenderingFunction} from './stencil-item-list-common.js';
 
 interface TableColumnsProps {
