+++ packages/atomic/src/components/ipx/atomic-ipx-result-link/atomic-ipx-result-link.tsx
@@ -12,7 +12,7 @@ import {
 } from '../../../utils/initialization-utils';
 import {buildStringTemplateFromResult} from '../../../utils/result-utils';
 import {getDefaultSlotFromHost} from '../../../utils/slot-utils';
-import {AnyUnfoldedItem} from '../../common/interface/item';
+import {AnyUnfoldedItem} from '../../common/item-list/unfolded-item';
 import {getAttributesFromLinkSlot} from '../../common/item-link/attributes-slot';
 import {LinkWithItemAnalytics} from '../../common/item-link/stencil-item-link';
 import {RecsBindings} from '../../recommendations/atomic-recs-interface/atomic-recs-interface';
