+++ packages/atomic/src/components/search/result-lists/atomic-folded-result-list/atomic-folded-result-list.tsx
@@ -33,7 +33,7 @@ import {FocusTargetController} from '../../../../utils/stencil-accessibility-uti
 import {shouldDisplayOnCurrentTab} from '../../../../utils/tab-utils';
 import {randomID} from '../../../../utils/utils';
 import {ResultsPlaceholdersGuard} from '../../../common/atomic-result-placeholder/stencil-placeholders';
-import {extractUnfoldedItem} from '../../../common/interface/item';
+import {extractUnfoldedItem} from '../../../common/item-list/unfolded-item';
 import {createAppLoadedListener} from '../../../common/interface/store';
 import {ItemDisplayGuard} from '../../../common/item-list/item-display-guard';
 import {ItemListGuard} from '../../../common/item-list/item-list-guard';
