+++ packages/atomic/src/components/search/result-lists/atomic-result-children/atomic-result-children.tsx
@@ -19,7 +19,7 @@ import {
   InitializeBindings,
 } from '../../../../utils/initialization-utils';
 import {elementHasAncestorTag} from '../../../../utils/utils';
-import {extractUnfoldedItem} from '../../../common/interface/item';
+import {extractUnfoldedItem} from '../../../common/item-list/unfolded-item';
 import {
   FoldedItemListContext,
   FoldedItemListStateContext,
